% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_all_chat_format}
\alias{check_all_chat_format}
\title{Check the chat format of all given files}
\usage{
check_all_chat_format(files)
}
\arguments{
\item{files}{A vector of file paths to Zoom chat files}
}
\value{
A two column tibble with columns \code{file} (the file path) and \code{type} ("One Liner" or "Two Liner")
}
\description{
Are these Zoom chat files in one-line (older) or two-line (newer) format?
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
