% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_chat_format}
\alias{check_chat_format}
\title{Check chat format of a single file}
\usage{
check_chat_format(file)
}
\arguments{
\item{file}{File path to a Zoom chat file}
}
\value{
Character vector, "v1" if in one-line format, "v2" if in two-line format
}
\description{
Is the Zoom chat file in one-line (older) or two-line (newer) format?
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
