% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_in_out}
\alias{check_in_out}
\title{Check if in and out folders exist}
\usage{
check_in_out(indir, outdir)
}
\arguments{
\item{indir}{Input directory path}

\item{outdir}{Output directory path}
}
\value{
Stops if they don't
}
\description{
Support function for reading and writing files
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
