% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_fdates}
\alias{get_fdates}
\title{Get file dates}
\usage{
get_fdates(fnames)
}
\arguments{
\item{fnames}{Vector of Zoom Chat file names}
}
\value{
Vector of YYYY-MM-DD dates the length of \code{fnames}
}
\description{
Extract YYYY-MM-DD dates from file names
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
