% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_zoom_names}
\alias{get_zoom_names}
\title{Get a vector of Zoom chat files}
\usage{
get_zoom_names(
  path = "~/Documents/Zoom",
  regexp = "meeting_saved_chat.txt|chat.txt",
  recurse = TRUE,
  ...
)
}
\arguments{
\item{path}{The director to look in, defaults to ~/Documents/Zoom}

\item{regexp}{The regexp to search for}

\item{recurse}{Look in all the subdirectories? Defaults to TRUE}

\item{...}{Other arguments passed to fs::dir_ls}
}
\value{

}
\description{
Point at directory, get chat txt file paths and names
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
