% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parse_all_v2_chat}
\alias{parse_all_v2_chat}
\title{Parse a series of version 2 Zoom chat files}
\usage{
parse_all_v2_chat(files)
}
\arguments{
\item{files}{A vector of version 2 Zoom chat files}
}
\value{
A tibble of parsed chats with columns \code{file} ((chr) source file name), \code{time} ((dttm) date and time of chat), \code{from} ((chr) From and To), \code{message} ((chr) Message content)
}
\description{
Parse a series of version 2 Zoom chat files
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
