% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parse_all_zoom_chat}
\alias{parse_all_zoom_chat}
\title{Parse a series of Zoom chats}
\usage{
parse_all_zoom_chat(files)
}
\arguments{
\item{files}{A vector of file paths to Zoom chat files}
}
\value{
A tibble of parsed chats with columns \code{file} ((chr) source file name), \code{time} ((dttm) date and time of chat), \code{from} ((chr) From and To), \code{message} ((chr) Message content)
}
\description{
Turn a vector of file paths to Zoom chat files into a tibble of chats with one chat message per row.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
