% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parse_v1_chat}
\alias{parse_v1_chat}
\title{Parse a Version 1 chat file into a tibble}
\usage{
parse_v1_chat(file)
}
\arguments{
\item{file}{Path to a version 1 Zoom chat file}
}
\value{
A tibble with columns \code{time} (dttm), \code{from} (chr), \code{to} (chr), and \code{message} (chr)
}
\description{
Parse a Version 1 chat file into a tibble
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
