% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parse_v2_chat}
\alias{parse_v2_chat}
\title{Parse a version 2 chat file}
\usage{
parse_v2_chat(file)
}
\arguments{
\item{file}{File path for a version 2 Zoom chat file}
}
\value{
A tibble of the chat with columns \code{time} (dttm), \code{from} (chr), \code{to} (chr), and \code{message} (chr)
}
\description{
Parse a version 2 chat file
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
