% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parse_zoom_chat}
\alias{parse_zoom_chat}
\title{Parse a Zoom Chat file}
\usage{
parse_zoom_chat(file)
}
\arguments{
\item{file}{Path to a Zoom chat file (v1 or v2)}
}
\value{
A tibble of the chat with columns \code{time} (dttm), \code{from} (chr), \code{to} (chr), and \code{message} (chr)
}
\description{
Turn a Zoom chat file into a tibble with one chat message per row
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
