% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{process_from_column}
\alias{process_from_column}
\title{Process from column}
\usage{
process_from_column(df)
}
\arguments{
\item{df}{A tibble of chat messages}
}
\value{
A tibble with from and to columns split out
}
\description{
Split and clean the from column into from and to components
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
